Fix select statement in metrics.q to work with kdb 4 (#51)
Select statement in upd function breaks in kdb 4.0, as it selects a column col
that is also used in the by clause. Removing the column in the select part of
the statement leads to the code giving the same result, so is redundant and
removed, allowing upd function to work in kdb 4.0.